Don't include used nodes in ready nodes number
nodepool by default includes used nodes in the ready count when
calculating the number of nodes that need to be spawned. Removing
this lets nodepool be a more aggressive by spawning nodes as soon
as a node with a given label goes used, rather than waiting for that
node to be offline/deleted before spawning a new one.

This works in conjunction with the gearman integration to allow
us to spawn nodes on-demand, even for labels with min-ready of 0.